lesson15 kadai1

+++ app/Http/Controllers/TasksController.php
@@ -1,34 +1,52 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
-
 use App\Task;    // 追加
+
 
 class TasksController extends Controller
 {
+
+    
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
-    {
-         // タスク一覧を取得
-        $tasks = Task::all();
-
-        // メッセージ一覧ビューでそれを表示
+    {   
+        if (\Auth::check()) { // 認証済みの場合
+        
+         // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
+        $data = [
+                'tasks' => $tasks,
+            ];
+        
+        // indexビューでそれらを表示
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
+        }
+        // ログインしていないときの処理
+        return view('auth.login');
     }
-
+    
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
-    {
+    {   
+        if (\Auth::check()) { // 認証済みの場合
+        
         $task = new task;
 
         // タスク作成ビューを表示
         return view('tasks.create', [
             'task' => $task,
         ]);
+        } else{
+        return view('auth.login');
+        }
     }
 
     // postでtasks/にアクセスされた場合の「新規登録処理」
@@ -40,12 +58,12 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        
-        // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            
+        'status' => $request->status,   // 追加
+        'content' => $request->content,
+        ]);
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -99,10 +117,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        $task = \App\Task::findOrFail($id);
+        
+     // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
